WCVP Taxo: append self-referenced data to backbone
In the new export, the parent taxon and the accepted name for a synonym
is no longer provided explicitly as another column (same for the
associated authors and POWO Id). Instead, the species name, name authors
and POWO Id must be fetched by self-referencing the table using the
plant_name_id, acting as a primary key, and the Ids acting as foreign
keys (parent_plant_name_id and accepted_plant_name_id).

This change is needed to convert the new WCVP export format to the old
format, and so to avoid further modifications in this scipt (it acts as
a reconciliation layer).