WCVP Taxo: rename backbone columns to reconcile format
As by the previous commit, another aspect to reconcile in the WCVP
backbone used as an input here, is to rename the Ids. Before, "Kew Id"
was used, but now that denomination doesn't exists: it is either IPNI Id
or POWO Id. Between IPNI and POWO Id, I have chosen POWO Id because it
is the Id we use in the database (POWO Id includes some temporary Ids
for the accepted names that haven't been already added to IPNI, and so
don't have an IPNI Id assigned yet, but that can be used to find the
name in POWO).