[barcode recovery] Use alternative fastq path
Use path valid in different hosts. Both paths cannot be valid
simultaneously. The path added points to the symlinks to all the SRA
fastq files downloaded in the data mining phase of PAFTOL 2.

Ideally, the path would be provided as an argument, but this is to avoid
modifying and testing now the script downstream (GetOrg_Pipeline.sh).