[DB] Download the latest BOLD barcodes dataset
Since 2021, there is a newer version of the BOLD's API (v4).  In the
latest version, one can't filter by class, but it is possible to
retrieve the data by taxonomic order. This is the reason for a
programmatic download that handles the download of the datasets relative
to each Angiosperm's order independently. The program then merges them
together in a single file.